backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-multipart==0.0.6
biopython==1.81
numpy==1.26.2
pandas==2.1.3
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
